# openapi.yml
openapi: 3.0.0
info:
  title: Perun Proxy API
  description: Perun Proxy API
  version: 0.3.0-dev
servers:
  - url: 'https://perun.cesnet.cz/rest/v1'
components:
  securitySchemes:
    BasicAuth:
      type: http
      scheme: basic
...
  responses:
    400BadRequest:
      description: Bad Request
    401NotAuthorized:
      description: Not authorized
    404NotFound:
      description: Not found
    405NotAllowed:
      description: Method not allowed
    500InternalError:
      description: Internal error
    503TemporaryUnavailable:
      description: Temporary unavailable
  examples:
    rp-identifier1:
      value: 'https://idp.example.com/idp/'
      summary: 'entityId'
    rp-identifier2:
      value: '123-456-879'
      summary: 'clientId'
    proxyuser-login:
      value: '1a2b3c@example.com'
      summary: User's login
paths:
...
  /auth/relying-party/{rp-identifier}/proxy-user/{login}/entitlements:
    get:
      summary: Get user entitlements
      parameters:
        - name: rp-identifier
          in: path
          required: true
          description: Relying party identifier in base64url safe format
          examples:
            entityId:
              $ref: '#/components/examples/rp-identifier1'
            clientId:
              $ref: '#/components/examples/rp-identifier2'
          schema:
            type: string
        - name: login
          in: path
          required: true
          description: User's login
          examples:
            login:
              $ref: '#/components/examples/proxyuser-login'
          schema:
            type: string
      security:
        - BasicAuth: []
      responses:
        '200':
          description: Returns list of user's entitlements.
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string
        '400':
          $ref: '#/components/responses/400BadRequest'
        '401':
          $ref: '#/components/responses/401NotAuthorized'
        '404':
          $ref: '#/components/responses/404NotFound'
        '405':
          $ref: '#/components/responses/405NotAllowed'
        '500':
          $ref: '#/components/responses/500InternalError'
        '503':
          $ref: '#/components/responses/503TemporaryUnavailable'
  /auth/relying-party/{rp-identifier}/proxy-user/{login}/entitlementsExtended:
    get:
      summary: Get user extended entitlements
      parameters:
        - name: rp-identifier
          in: path
          required: true
          description: Relying party identifier in base64url safe format
          examples:
            entityId:
              $ref: '#/components/examples/rp-identifier1'
            clientId:
              $ref: '#/components/examples/rp-identifier2'
          schema:
            type: string
        - name: login
          in: path
          required: true
          description: User's login
          examples:
            login:
              $ref: '#/components/examples/proxyuser-login'
          schema:
            type: string
      security:
        - BasicAuth: [ ]
      responses:
        '200':
          description: Returns list of user's extended entitlements.
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string
        '400':
          $ref: '#/components/responses/400BadRequest'
        '401':
          $ref: '#/components/responses/401NotAuthorized'
        '404':
          $ref: '#/components/responses/404NotFound'
        '405':
          $ref: '#/components/responses/405NotAllowed'
        '500':
          $ref: '#/components/responses/500InternalError'
        '503':
          $ref: '#/components/responses/503TemporaryUnavailable'
  /auth/relying-party/{rp-identifier}/proxy-user/{login}/access:
    get:
      summary: Check if user has access to client
      parameters:
        - name: rp-identifier
          in: path
          required: true
          description: Relying party identifier in base64url safe format
          examples:
            entityId:
              $ref: '#/components/examples/rp-identifier1'
            clientId:
              $ref: '#/components/examples/rp-identifier2'
          schema:
            type: string
        - name: login
          in: path
          required: true
          description: User's login
          example: '1a2b3c@example.com'
          schema:
            type: string
      security:
        - BasicAuth: []
      responses:
        '200':
          description: Returns TRUE if user has access to service, otherwise FALSE.
          content:
            application/json:
              schema:
                type: boolean
        '400':
          $ref: '#/components/responses/400BadRequest'
        '401':
          $ref: '#/components/responses/401NotAuthorized'
        '404':
          $ref: '#/components/responses/404NotFound'
        '405':
          $ref: '#/components/responses/405NotAllowed'
        '500':
          $ref: '#/components/responses/500InternalError'
        '503':
          $ref: '#/components/responses/503TemporaryUnavailable'
  /auth/relying-party/{rp-identifier}/environment:
    get:
      summary: Get environment of the RP
      parameters:
        - name: rp-identifier
          in: path
          required: true
          description: Relying party identifier in base64url safe format
          examples:
            entityId:
              $ref: '#/components/examples/rp-identifier1'
            clientId:
              $ref: '#/components/examples/rp-identifier2'
          schema:
            type: string
      security:
        - BasicAuth: [ ]
      responses:
        '200':
          description: Returns one of STAGING, PRODUCTION, TESTING
          content:
            application/json:
              schema:
                type: string
        '400':
          $ref: '#/components/responses/400BadRequest'
        '401':
          $ref: '#/components/responses/401NotAuthorized'
        '404':
          $ref: '#/components/responses/404NotFound'
        '405':
          $ref: '#/components/responses/405NotAllowed'
        '500':
          $ref: '#/components/responses/500InternalError'
        '503':
          $ref: '#/components/responses/503TemporaryUnavailable'
...